requirements.txt
python-twitter==3.3
sklearn
colorama
keyboard
numpy
scipy
click
liac-arff
python3-weka-wrapper3
yagmail